Fix tmux.sh and use /mnt/ for synced_folder

+++ tmux.sh
@@ -1,15 +1,18 @@
 #!/bin/bash
 
-curl https://github.com/tmux/tmux/releases/download/2.3/tmux-2.3.tar.gz -O -L
-tar -xf tmux-2.3.tar.gz
-rm tmux-2.3.tar.gz
-cd tmux-2.3
+if test "$(tmux -V || echo 'no tmux')" != "tmux 2.3"; then
+  curl https://github.com/tmux/tmux/releases/download/2.3/tmux-2.3.tar.gz -O -L
+  tar -xf tmux-2.3.tar.gz
+  rm tmux-2.3.tar.gz
+  cd tmux-2.3
 
-./configure && make
+  ./configure && make
 
-# Install as root
-make install
+  # Install as root
+  make install
 
-# Cleanup
-cd ../
-rm -rf tmux-2.3
+  # Cleanup
+  cd ../
+  rm -rf tmux-2.3
+else echo "tmux already installed";
+fi
